requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
blis==0.7.11
catalogue==2.0.10
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
distro==1.9.0
Django==5.2.12
djangorestframework==3.16.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.3.1
filelock==3.25.0
fsspec==2026.2.0
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
IMAPClient==3.1.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kombu==5.6.2
langcodes==3.5.1
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.4.2
numpy==1.26.4
packaging==26.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
preshed==3.0.12
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
pyzmail36==1.0.5
redis==7.3.0
regex==2026.2.28
requests==2.32.5
safetensors==0.7.0
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.3
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.2
sympy==1.14.0
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.9.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.3.4
whitenoise==6.12.0
